Extract st.write so it can be called from st.sidebar
